+++ test/nut/scratchDelete.nut.ts
