+++ test/shared/scratchOrgOutput.test.ts
