+++ test/unit/org/display.test.ts
@@ -8,9 +8,9 @@ import { expect, config as chaiConfig } from 'chai';
 import { MockTestOrgData, TestContext } from '@salesforce/core/lib/testSetup';
 import { Config } from '@oclif/core';
 import { Connection } from '@salesforce/core';
-import { OrgDisplayReturn } from '../../../../src/shared/orgTypes';
-import { OrgListUtil } from '../../../../src/shared/orgListUtil';
-import { OrgDisplayCommand } from '../../../../src/commands/org/display';
+import { OrgDisplayReturn } from '../../../src/shared/orgTypes';
+import { OrgListUtil } from '../../../src/shared/orgListUtil';
+import { OrgDisplayCommand } from '../../../src/commands/org/display';
 
 chaiConfig.truncateThreshold = 0;
