+++ test/unit/org/open.test.ts
@@ -6,16 +6,12 @@
  */
 
 import { assert, expect } from 'chai';
-import {
-  // Org,
-  MyDomainResolver,
-  Messages,
-} from '@salesforce/core';
+import { MyDomainResolver, Messages } from '@salesforce/core';
 import { Config } from '@oclif/core';
 import { stubMethod } from '@salesforce/ts-sinon';
 import { MockTestOrgData, TestContext, shouldThrow } from '@salesforce/core/lib/testSetup';
-import * as utils from '../../../../src/shared/utils';
-import { OrgOpenCommand, OrgOpenOutput } from '../../../../src/commands/org/open';
+import * as utils from '../../../src/shared/utils';
+import { OrgOpenCommand, OrgOpenOutput } from '../../../src/commands/org/open';
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('@salesforce/plugin-org', 'open');
